+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- NPM_TOKEN enabled for CI/CD.
+- Optimize svg files using svgo from script.
 
 ## 0.0.2 - 2024-02-16
 ### Added